+++ R/author-poiscon.r
@@ -1,9 +1,3 @@
-#' @title Author poiscon
-#'
-#' @description
-#' Creates slidify deck
-#' 
-#' @param dir string of github repository name.
 author_poiscon <- function (deck_dir = "output/deck",
                              jekyll_dir = "poissonconsulting.github.io", 
                              jekyll_path = "~/Documents/code") {
@@ -11,12 +5,12 @@ author_poiscon <- function (deck_dir = "output/deck",
   if(!file.exists(jekyll_path))
     stop("jekyll_path does not exist")
   
-  if(!file.exists(paste0(path,"/",dir)))
+  if(!file.exists(paste0(jekyll_path,"/",deck_dir)))
     stop("jekyll_dir does not exist")
   
-  if(!file.exists(paste0(path,"/",dir,"/-posts")))
+  if(!file.exists(paste0(jekyll_path,"/",jekyll_dir,"/-posts")))
     stop("jekyll_dir is not a jekyll repository")
   
-  return (paste0(path,"/",dir))
+  return (paste0(jekyll_path,"/",jekyll_dir))
   
 }