+++ R/figures-web.r
@@ -5,10 +5,16 @@
 #' output/report/project-name/figures ready for linking to the report and
 #' posting on the jekyll site in the figures directory.
 #' 
+#' @param dir jekyll repository
 #' @return An invisible logical scalar indicating whether or not sucessful.
 #' @export
-extract_figures <- function () {
+figures_web <- function (dir = "poissonconsulting.github.io") {
   
+  assert_that(is.string(dir))
+  assert_that(is.string(options()$code_dir))
+  
+  path <- options()$code_dir
+
   from <- "output/plots/figures"
   
   files <- list.files(from, pattern = "[.]png", full.names = TRUE, 
@@ -16,19 +22,18 @@ extract_figures <- function () {
   
   if(length(files)) {
     
-    to <- str_replace_all(files, "plots/figures", 
-                          paste0("report/",project_folder(),"/figures"))
+    to <- str_replace_all(files, 
+                          "output/plots/figures",
+                          paste(path, dir, "figures", project_folder(), 
+                                sep = "/"))
     
     folders <- str_replace_all(to, "[/][^/]*[.][p][n][g]$", "")
     folders <- unique(folders)
     
     for (folder in folders) 
       dir.create(folder, recursive = TRUE, showWarnings = FALSE)
-    
+            
     file.copy(files, to, overwrite = TRUE)
-    file.copy(str_replace_all(files,"[.]png",".csv"), 
-              str_replace_all(to,"[.]png",".csv"), overwrite = TRUE)
-    
   }
   return (invisible())
 }