+++ R/knit-report.r
@@ -3,7 +3,7 @@
 #' @description
 #' Knits report
 #' 
-#' @param web_dir a character scalar or a Date
+#' @param file a character scalar of file
 #' @return Knits readme file into markdown
 #' @export
 knit_report <- function (file = "report.rmd") {
@@ -11,28 +11,71 @@ knit_report <- function (file = "report.rmd") {
   assert_that(is.string(file))
   assert_that(file.exists(file))
   
+  figures_report()
+  
   to <- paste0("output/report/", project_folder(), "/report.md")
   
   if(!file.exists(str_replace(to, "/report.md", "")))
     dir.create(str_replace(to, "/report.md", ""), FALSE, recursive = TRUE)
   
   knit(file, output = to)
-  
-  from <- to
-  to <- str_replace(to, "report.md", "temporary.md")
     
+  from <- to
+  to_html <- str_replace(to, "report.md", "html.md")
+  to_docx <- str_replace(to, "report.md", "docx.md")
+  rm(to)
+  
   rm_blank_lines_top(from)
   str_replace_file(from, "&", "and")
   
-  str_replace_file(from, 
-                   "[#][#].[B][a][c][k][g][r][o][u][n][d]", 
-                   paste0("#### ", Sys.Date(), "\n## Background"), to = to)
+  title <- header_setting(from, "title")
+  layout <- header_setting(from, "layout")
   
-  markdownToHTML(to, str_replace(from, "*[.][m][d]$", ".html"))
+  stopifnot(layout %in% c("page", "post"))
+
+  
+  if(layout == "page") {
+    str_replace_file(from, 
+                     "[#][#][ ][B][a][c][k][g][r][o][u][n][d]", 
+                     paste0("# ", title, "\n### Draft: ", Sys.Date(), 
+                            "\n## Background"),
+                     to = to_html)  
     
-  system(paste0("pandoc -s ", to," -o ", str_replace(from,"[.][m][d]$",".docx")))
+    str_replace_file(from, 
+                     "[#][#][ ][B][a][c][k][g][r][o][u][n][d]", 
+                     paste0("### Draft: ", Sys.Date(), "\n## Background"),
+                     to = to_docx)    
+    
+    str_replace_file(from, 
+                     "[#][#][ ][B][a][c][k][g][r][o][u][n][d]", 
+                     paste0("### Draft: ", Sys.Date(), "\n## Background"))      
+  } else {
+    date <- as.Date(header_setting(from, "release_date"))
+    
+    stopifnot(is.Date(date))
+    
+    str_replace_file(from, 
+                     "[{][%][ ][i][n][c][l][u][d][e][ ][J][B][/][s][e][t][u][p][ ][%]}",
+                     "",
+                     to = to_docx)
 
-  file.remove(to)
+    str_replace_file(from, 
+                     "[{][%][ ][i][n][c][l][u][d][e][ ][J][B][/][s][e][t][u][p][ ][%]}",
+                     paste("#", title, "\n###", date),
+                     to = to_html)  
 
+    str_replace_file(from, 
+                     "[{][%][ ][i][n][c][l][u][d][e][ ][J][B][/][s][e][t][u][p][ ][%]}",
+                     paste("###", date),
+                     to = to_docx)
+  }
+  
+  markdownToHTML(to_html, str_replace(from, "*[.][m][d]$", ".html"))
+    
+  system(paste0("pandoc -s ", to_docx," -o ", str_replace(from,"[.][m][d]$",".docx")))
+  
+  file.remove(to_html)
+  file.remove(to_docx)
+  
   return (invisible())
 }